use ordered sets to hold attribute values
According to the LDAP RFCs, attribute value order doesn't matter.
However, OpenLDAP uses the X-ORDERED extension specified in
draft-chu-ldap-xordered-00, and order does matter for that extension.
Use the WeaklyOrderedSet class instead of set to hold attribute
values.  This causes the attribute values to be given to the server in
the order given by the user, assuming the backend preserves the order
when generating the ASN.1.

Fixes issue #33815.